Re-namespace project529 classes
